Remove all our REAL KIND hacks
And use the DisableType mechanism to disable only KIND=2 and KIND=3

Also make the disable a hard error to avoid funky problems.

+++ flang/test/Lower/Intrinsics/rrspacing.f90
@@ -2,10 +2,10 @@
 ! RUN: %flang_fc1 -emit-fir %s -o - | FileCheck %s
 
 ! CHECK-LABEL: func @_QPrrspacing_test2(
-! CHECK-SAME: %[[x:[^:]+]]: !fir.ref<f64>{{.*}}) -> f64
-real*8 function rrspacing_test2(x)
-  real*8 :: x
+! CHECK-SAME: %[[x:[^:]+]]: !fir.ref<f128>{{.*}}) -> f128
+real*16 function rrspacing_test2(x)
+  real*16 :: x
   rrspacing_test2 = rrspacing(x)
-! CHECK: %[[a1:.*]] = fir.load %[[x]] : !fir.ref<f64>
-! CHECK: %{{.*}} = fir.call @_FortranARRSpacing8(%[[a1]]) : (f64) -> f64
+! CHECK: %[[a1:.*]] = fir.load %[[x]] : !fir.ref<f128>
+! CHECK: %{{.*}} = fir.call @_FortranARRSpacing16(%[[a1]]) : (f128) -> f128
 end function
